copy-paste simd in cpu
